requirements.txt
arrow==0.13.1
asn1crypto==0.24.0
attrdict==2.0.1
certifi==2019.3.9
cffi==1.12.2
chardet==3.0.4
coincurve==11.0.0
cytoolz==0.9.0.1
Django==2.1.7
ecdsa==0.13
eth-abi==1.3.0
eth-account==0.3.0
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.2.1
eth-rlp==0.1.2
eth-typing==2.1.0
eth-utils==1.4.1
ethereum==2.3.2
future==0.17.1
hexbytes==0.1.0
idna==2.8
lru-dict==1.1.6
mypy-extensions==0.4.1
parsimonious==0.8.1
pbkdf2==1.3
py-ecc==1.6.0
pycparser==2.19
pycryptodome==3.7.3
pycryptodomex==3.7.3
pyethash==0.1.27
pysha3==1.0.2
python-dateutil==2.8.0
pytz==2018.9
PyYAML==5.1
repoze.lru==0.7
requests==2.21.0
rlp==1.1.0
scrypt==0.8.13
simplejson==3.16.0
six==1.12.0
toolz==0.9.0
urllib3==1.24.1
web3==4.8.3
websockets==6.0
